pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fx-notifier"
version = "0.1.0"
description = "Fetch FX rates from Frankfurter and send Telegram notifications."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.0",
  "python-telegram-bot>=20.0",
  "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.11.0",
  "pytest>=8.0.0",
  "ruff>=0.11.0",
]

[project.scripts]
fx-notifier = "fx_notifier.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
fx_notifier = ["py.typed"]

[tool.pytest.ini_options]
addopts = "-q -p no:cacheprovider"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["fx_notifier"]

[tool.mypy]
python_version = "3.10"
files = ["src"]
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unused_configs = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = ["requests", "telegram", "telegram.*"]
ignore_missing_imports = true
